Linked placeOrder page to company page
