+++ front-end/src/components/CompanyDetail/CompanyActions.js
@@ -3,14 +3,26 @@ import PropTypes from "prop-types";
 import clsx from "clsx";
 import { isEqual, isEmpty, pick } from "lodash";
 import { withRouter } from "react-router";
+import { socket } from "../../App";
 
 import { connect } from "react-redux";
+import { orderAction, userAction } from "../../redux/storeActions/actions";
+
+import { redirectToPage } from "../../utils/low-dependency/PageRedirectUtil";
+import { changeUserData } from "../../utils/UserUtil";
 
 import { withStyles } from "@material-ui/core/styles";
-import { Paper, Typography, Divider } from "@material-ui/core";
+import {
+  Paper,
+  Typography,
+  Divider,
+  Button,
+  ButtonGroup
+} from "@material-ui/core";
 
 const styles = (theme) => ({
   actionsPaper: {
+    width: "100%",
     padding: "20px",
     backgroundColor: theme.palette.paperBackground.onPage,
   },
@@ -24,7 +36,16 @@ const styles = (theme) => ({
   divider: {
     backgroundColor: "rgba(255, 255, 255, 0.2)",
     marginTop: "10px",
-    marginBottom: "10px",
+    marginBottom: "18px",
+  },
+  actionButton: {
+    color: theme.palette.primary.main,
+    border: `1px solid ${theme.palette.primary.main}`,
+    fontWeight: "bold",
+  },
+  watchlistButtonRemove: {
+    color: theme.palette.secondary.main,
+    border: `1px solid ${theme.palette.secondary.main}`,
   },
 });
 
@@ -35,6 +56,48 @@ class CompanyActionsPaper extends React.Component {
     const propsCompare = pick(this.props, compareKeys);
     return !isEqual(nextPropsCompare, propsCompare);
   }
+
+  handlePlaceOrder = () => {
+    this.props.mutateOrder({companyCode: this.props.companyCode});
+    redirectToPage('/placeOrder', this.props)
+  };
+
+  addToWatchlist = () => {
+    const { companyCode } = this.props;
+    let { email, watchlist: newWatchlist } = this.props.userSession;
+
+    newWatchlist.push(companyCode);
+    const dataNeedChange = {
+      watchlist: {
+        set: newWatchlist,
+      },
+    };
+    changeUserData(dataNeedChange, email, this.props.mutateUser, socket).catch(
+      (err) => {
+        console.log(err);
+      }
+    );
+  };
+
+  removeFromWatchlist = () => {
+    const { companyCode } = this.props;
+    let { email, watchlist: newWatchlist } = this.props.userSession;
+
+    newWatchlist = newWatchlist.filter(
+      (companyCodeString) => companyCodeString !== companyCode
+    );
+
+    const dataNeedChange = {
+      watchlist: {
+        set: newWatchlist,
+      },
+    };
+    changeUserData(dataNeedChange, email, this.props.mutateUser, socket).catch(
+      (err) => {
+        console.log(err);
+      }
+    );
+  };
 
   render() {
     const {
@@ -48,7 +111,7 @@ class CompanyActionsPaper extends React.Component {
     return (
       <Paper className={clsx(classes.actionsPaper, paperClassName)}>
         <Typography className={clsx(classes.title, titleClassName)}>
-          {`Buy / Sell ${companyCode}`}
+          {`Actions for ${companyCode}`}
         </Typography>
 
         <Divider className={classes.divider} />
@@ -59,9 +122,38 @@ class CompanyActionsPaper extends React.Component {
           </Typography>
         )}
 
-        {!isEmpty(userSession) && (
-          <Typography className={classes.body}>Let's buy stock!</Typography>
-        )}
+        <ButtonGroup
+          orientation="vertical"
+          aria-label="action button group"
+          fullWidth
+        >
+          <Button
+            onClick={this.handlePlaceOrder}
+            aria-label="trade button"
+            className={classes.actionButton}
+          >
+            Trade
+          </Button>
+
+          <Button
+            aria-label="add/remove watchlist button"
+            className={clsx(classes.actionButton, {
+              [classes.watchlistButtonRemove]: userSession.watchlist.includes(
+                companyCode
+              ),
+            })}
+            onClick={
+              userSession.watchlist.includes(companyCode)
+                ? this.removeFromWatchlist
+                : this.addToWatchlist
+            }
+          >
+            {userSession.watchlist.includes(companyCode) &&
+              "Remove from watchlist"}
+            {!userSession.watchlist.includes(companyCode) &&
+              "Add to watchlist"}
+          </Button>
+        </ButtonGroup>
       </Paper>
     );
   }
@@ -78,6 +170,12 @@ const mapStateToProps = (state) => ({
   userSession: state.userSession,
 });
 
-export default connect(mapStateToProps)(
-  withStyles(styles)(withRouter(CompanyActionsPaper))
-);
+const mapDispatchToProps = (dispatch) => ({
+  mutateUser: (userProps) => dispatch(userAction("default", userProps)),
+  mutateOrder: (dataToChange) => dispatch(orderAction("change", dataToChange)),
+});
+
+export default connect(
+  mapStateToProps,
+  mapDispatchToProps
+)(withStyles(styles)(withRouter(CompanyActionsPaper)));
